+++ apps/frontend/index.html
@@ -13,7 +13,7 @@
     />
   </head>
   <body>
-    <div id="app"></div>
+    <div id="app" class="font-body relative min-h-screen bg-stone-50 text-stone-800"></div>
     <script type="module" src="/src/app/main.ts"></script>
   </body>
 </html>
