+++ apps/frontend/src/app/router.ts
@@ -1,15 +1,13 @@
 import { createRouter, createWebHistory } from "vue-router";
-import Home from "@/pages/Home.vue";
+import { handleHotUpdate, routes } from "vue-router/auto-routes";
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
-  routes: [
-    {
-      path: "/",
-      name: "home",
-      component: Home,
-    },
-  ],
+  routes,
 });
+
+if (import.meta.hot) {
+  handleHotUpdate(router);
+}
 
 export default router;
