+++ apps/frontend/src/pages/_index/Hero.vue
