+++ apps/frontend/src/pages/_index/NewsletterCTA.vue
