+++ apps/frontend/src/pages/_index/TodaysPick.vue
