+++ apps/frontend/src/pages/_index/_hero/HeroRecipe.vue
