+++ apps/frontend/src/pages/_index/_hero/SocialProof.vue
