+++ apps/frontend/src/pages/_index/featured-recipes/FeaturedRecipe.vue
