+++ apps/frontend/src/pages/_index/featured-recipes/FeaturedRecipeFooter.vue
