+++ apps/frontend/src/pages/_index/featured-recipes/FeaturedRecipeHeader.vue
