+++ apps/frontend/tsconfig.app.json
@@ -1,18 +1,31 @@
 {
-  "extends": "@vue/tsconfig/tsconfig.dom.json",
-  "include": ["env.d.ts", "src/**/*", "src/**/*.vue"],
-  "exclude": ["src/**/__tests__/*"],
-  "compilerOptions": {
-    // Extra safety for array and object lookups, but may have false positives.
-    "noUncheckedIndexedAccess": true,
-
-    // Path mapping for cleaner imports.
-    "paths": {
-      "@/*": ["./src/*"]
+    "extends": "@vue/tsconfig/tsconfig.dom.json",
+    "include": [
+        "env.d.ts",
+        "./typed-router.d.ts",
+        "src/**/*",
+        "src/**/*.vue"
+    ],
+    "exclude": [
+        "src/**/__tests__/*"
+    ],
+    "compilerOptions": {
+        // Extra safety for array and object lookups, but may have false positives.
+        "noUncheckedIndexedAccess": true,
+        // Path mapping for cleaner imports.
+        "paths": {
+            "@/*": [
+                "./src/*"
+            ]
+        },
+        // `vue-tsc --build` produces a .tsbuildinfo file for incremental type-checking.
+        // Specified here to keep it out of the root directory.
+        "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo"
     },
-
-    // `vue-tsc --build` produces a .tsbuildinfo file for incremental type-checking.
-    // Specified here to keep it out of the root directory.
-    "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo"
-  }
+    "vueCompilerOptions": {
+        "plugins": [
+            "vue-router/volar/sfc-route-blocks",
+            "vue-router/volar/sfc-typed-router"
+        ]
+    }
 }
